requirements.txt

streamlit
pandas
seaborn
matplotlib
langchain
langchain-openai
langchain-experimental
openai
python-dotenv
